+++ src/tools/library.ts
@@ -17,10 +17,22 @@
  */
 
 import { z } from 'zod';
+import type { Tool } from '@modelcontextprotocol/sdk/types.js';
 import type { NavidromeClient } from '../client/navidrome-client.js';
+import type { Config } from '../config.js';
 import { transformSongsToDTO } from '../transformers/song-transformer.js';
 import type { SongDTO } from '../types/dto.js';
 import { DEFAULT_VALUES } from '../constants/defaults.js';
+import type { ToolCategory } from './handlers/registry.js';
+import {
+  listAlbums,
+  listArtists,
+  listGenres,
+  getSong,
+  getAlbum,
+  getArtist,
+  getSongPlaylists,
+} from './media-library.js';
 
 const ListSongsSchema = z.object({
   limit: z.number().min(1).max(500).optional().default(DEFAULT_VALUES.SONGS_LIMIT),
@@ -73,4 +85,230 @@ export async function listSongs(client: NavidromeClient, args: unknown): Promise
       `Failed to fetch songs: ${error instanceof Error ? error.message : 'Unknown error'}`
     );
   }
+}
+
+// Tool definitions for library category
+const tools: Tool[] = [
+  {
+    name: 'list_songs',
+    description: 'List songs from the Navidrome music library with clean, LLM-friendly data (filtering and pagination supported)',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        limit: {
+          type: 'number',
+          description: 'Maximum number of songs to return (1-500)',
+          minimum: 1,
+          maximum: 500,
+          default: DEFAULT_VALUES.SONGS_LIMIT,
+        },
+        offset: {
+          type: 'number',
+          description: 'Number of songs to skip for pagination',
+          minimum: 0,
+          default: 0,
+        },
+        sort: {
+          type: 'string',
+          description: 'Field to sort by',
+          enum: ['title', 'artist', 'album', 'year', 'duration', 'playCount', 'rating'],
+          default: 'title',
+        },
+        order: {
+          type: 'string',
+          description: 'Sort order',
+          enum: ['ASC', 'DESC'],
+          default: 'ASC',
+        },
+        starred: {
+          type: 'boolean',
+          description: 'Filter for starred songs only',
+        },
+      },
+    },
+  },
+  {
+    name: 'list_albums',
+    description: 'List albums from the Navidrome music library with clean, LLM-friendly data',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        limit: {
+          type: 'number',
+          description: 'Maximum number of albums to return (1-500)',
+          minimum: 1,
+          maximum: 500,
+          default: DEFAULT_VALUES.ALBUMS_LIMIT,
+        },
+        offset: {
+          type: 'number',
+          description: 'Number of albums to skip for pagination',
+          minimum: 0,
+          default: 0,
+        },
+        sort: {
+          type: 'string',
+          description: 'Field to sort by',
+          default: 'name',
+        },
+        order: {
+          type: 'string',
+          description: 'Sort order',
+          enum: ['ASC', 'DESC'],
+          default: 'ASC',
+        },
+      },
+    },
+  },
+  {
+    name: 'list_artists',
+    description: 'List artists from the Navidrome music library with clean, LLM-friendly data',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        limit: {
+          type: 'number',
+          description: 'Maximum number of artists to return (1-500)',
+          minimum: 1,
+          maximum: 500,
+          default: 100,
+        },
+        offset: {
+          type: 'number',
+          description: 'Number of artists to skip for pagination',
+          minimum: 0,
+          default: 0,
+        },
+        sort: {
+          type: 'string',
+          description: 'Field to sort by',
+          default: 'name',
+        },
+        order: {
+          type: 'string',
+          description: 'Sort order',
+          enum: ['ASC', 'DESC'],
+          default: 'ASC',
+        },
+      },
+    },
+  },
+  {
+    name: 'list_genres',
+    description: 'List all genres from the Navidrome music library',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        limit: {
+          type: 'number',
+          description: 'Maximum number of genres to return (1-500)',
+          minimum: 1,
+          maximum: 500,
+          default: 100,
+        },
+        offset: {
+          type: 'number',
+          description: 'Number of genres to skip for pagination',
+          minimum: 0,
+          default: 0,
+        },
+        sort: {
+          type: 'string',
+          description: 'Field to sort by',
+          default: 'name',
+        },
+        order: {
+          type: 'string',
+          description: 'Sort order',
+          enum: ['ASC', 'DESC'],
+          default: 'ASC',
+        },
+      },
+    },
+  },
+  {
+    name: 'get_song',
+    description: 'Get detailed information about a specific song by ID',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        id: {
+          type: 'string',
+          description: 'The unique ID of the song',
+        },
+      },
+      required: ['id'],
+    },
+  },
+  {
+    name: 'get_album',
+    description: 'Get detailed information about a specific album by ID',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        id: {
+          type: 'string',
+          description: 'The unique ID of the album',
+        },
+      },
+      required: ['id'],
+    },
+  },
+  {
+    name: 'get_artist',
+    description: 'Get detailed information about a specific artist by ID',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        id: {
+          type: 'string',
+          description: 'The unique ID of the artist',
+        },
+      },
+      required: ['id'],
+    },
+  },
+  {
+    name: 'get_song_playlists',
+    description: 'Get all playlists that contain a specific song',
+    inputSchema: {
+      type: 'object',
+      properties: {
+        songId: {
+          type: 'string',
+          description: 'The unique ID of the song',
+        },
+      },
+      required: ['songId'],
+    },
+  },
+];
+
+// Factory function for creating library tool category with dependencies  
+export function createLibraryToolCategory(client: NavidromeClient, config: Config): ToolCategory {
+  return {
+    tools,
+    async handleToolCall(name: string, args: unknown): Promise<unknown> {
+      switch (name) {
+        case 'list_songs':
+          return await listSongs(client, args);
+        case 'list_albums':
+          return await listAlbums(client, args);
+        case 'list_artists':
+          return await listArtists(client, args);
+        case 'list_genres':
+          return await listGenres(client, config, args);
+        case 'get_song':
+          return await getSong(client, args);
+        case 'get_album':
+          return await getAlbum(client, args);
+        case 'get_artist':
+          return await getArtist(client, args);
+        case 'get_song_playlists':
+          return await getSongPlaylists(client, args);
+        default:
+          throw new Error(`Unknown library tool: ${name}`);
+      }
+    }
+  };
 }
